fix(gateway): move resolve e2e harness to the sibling integration/http convention
Review follow-up on the GET /v1/items/resolve PR:
- Extract makeInMemoryVault into test/helpers/in-memory-vault.ts (was a near-verbatim
  third copy of agent-test-server.ts / brief-test-server.ts's local copies).
- Move startServerWithClipToken into src/ipc/http-api-test-server.ts, a non-.test.ts
  exported harness paralleling agent-test-server.ts / brief-test-server.ts. Adds
  startServerWithoutClipsVault for the surface-not-mounted case, and generalizes
  extraOpts pass-through so a later POST /v1/items/fetch task can reuse it.
- Move the assertions to test/integration/http/items-resolve-route.test.ts (matching
  the four sibling *-route.test.ts files) and delete test/http-api.test.ts, which never
  matched the established convention.
- Add the previously-untested 'surface not mounted' 404 case, pinned against the
  agent-http-e2e.test.ts:200-216 precedent.
